requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
pillow-heif==0.13.0
Pillow==10.1.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0